requirements.txt
asgiref==3.4.0
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-multiselectfield==0.1.12
et-xmlfile==1.1.0
idna==2.10
lml==0.1.0
MarkupPy==1.14
numpy==1.21.0
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.0
Pillow==8.3.1
pycodestyle==2.7.0
pycparser==2.20
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
sqlparser==0.0.9
tablib==3.0.0
texttable==1.6.4
toml==0.10.2
urllib3==1.26.6
waitress==2.0.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
